requirements.txt
PyGithub==1.2
clint==0.3.1
mockito==0.5.1
pymongo==2.2
pycrypto==2.6
